+++ apps/koku-ui-ros/src/routes/optimizations/optimizationsBreakdown/nodeBreakdown/visualInsights/index.ts
@@ -1,1 +1,2 @@
+export { NodeRequestGapChart } from './nodeRequestGapChart';
 export { NodeUtilizationTrend } from './nodeUtilizationTrend';